Linux fixes from Danni.

+++ Fuji/Source/Drivers/Sound/MFSound_OpenAL.cpp
@@ -311,7 +311,7 @@ void MFSound_DestroyInternal(MFSound *pSound)
 	}
 }
 
-MF_API int MFSound_SetBufferData(MFSound *pSound, const void *pData, uint32 size)
+MF_API size_t MFSound_SetBufferData(MFSound *pSound, const void *pData, size_t size)
 {
 	MFSoundTemplate *pTemplate = pSound->pTemplate;
 	MFSoundDataInternal *pInternal = pSound->pInternal;
@@ -401,7 +401,7 @@ MF_API int MFSound_SetBufferData(MFSound *pSound, const void *pData, uint32 size
 	return size;
 }
 
-MF_API int MFSound_Lock(MFSound *pSound, size_t offset, size_t bytes, void **ppData, uint32 *pSize, void **ppData2, uint32 *pSize2)
+MF_API int MFSound_Lock(MFSound *pSound, size_t offset, size_t bytes, void **ppData, size_t *pSize, void **ppData2, size_t *pSize2)
 {
 	MFDebug_Assert(!(pSound->flags & MFPF_Locked), MFStr("Dynamic sound '%s' is already locked.", pSound->name));
